+++ src/AwesomeApp.UI/src/backend/libs/ApiMiddlewareHandlers.ts
