+++ src/AwesomeApp.UI/src/backend/libs/index.ts
@@ -1,3 +1,3 @@
 export * from './ApiClient'
 export * from './IronSessionConfig'
-export * from './ApiHandlers'
+export * from './ApiMiddlewareHandlers'
