+++ src/AwesomeApp.UI/src/pages/api/account/authenticate.ts
@@ -1,6 +1,6 @@
 import { withAuthentication, withEndpoints, withErrorHandling } from "@/backend/libs";
 import { HttpMethod } from "@/shared/HttpMethod";
-import { authenticate } from "@/backend/controllers/AccountController";
+import { authenticate } from "@/backend/handlers/AccountHandlers";
 
 export default 
 withErrorHandling(
